Shooting now works

+++ Robotron/Robotron/EntityUtils.cpp
@@ -60,7 +60,7 @@ Entity& EntityUtils::createEnemyZombie(Scene& scene, const glm::mat4& transform)
 	entity.aiVariables.score = 5;
 	entity.aiVariables.lives = 1;
 
-	float pi = M_PI_2;
+	float pi = static_cast<float>(M_PI_2);
 	entity.transform = transform * glm::rotate(glm::mat4{}, pi, glm::vec3{ 1.0f, 0.0f, 0.0f });
 
 	entity.model = GLPrimitives::getSphereModel();
@@ -89,7 +89,7 @@ Entity& EntityUtils::createEnemySnake(Scene& scene, const glm::mat4& transform, 
 	entity.aiVariables.score = 5;
 	entity.aiVariables.lives = 1;
 
-	float pi = M_PI_2;
+	float pi = static_cast<float>(M_PI_2);
 	entity.transform = transform * glm::rotate(glm::mat4{}, pi, glm::vec3{ 1.0f, 0.0f, 0.0f });
 
 	entity.model = GLPrimitives::getSphereModel();
@@ -112,7 +112,7 @@ Entity& EntityUtils::createEnemyShooter(Scene& scene, const glm::mat4& transform
 	Entity& entity = scene.createEntity(COMPONENT_MODEL | COMPONENT_TRANSFORM 
 	                                  | COMPONENT_ENEMY_SHOOTER
 	                                  | COMPONENT_NETWORK | COMPONENT_PHYSICS);
-	float pi = M_PI_2;
+	float pi = static_cast<float>(M_PI_2);
 	entity.transform = transform * glm::rotate(glm::mat4{}, pi, glm::vec3{ 1.0f, 0.0f, 0.0f });
 
 	entity.aiVariables.score = 10;
